Init weld removal
