+++ execution-plane/arcon/Cargo.toml
@@ -22,7 +22,6 @@ arcon_messages = { version = "0.1", path = "arcon_messages" }
 arcon_backend = { version = "0.1", path = "arcon_backend" }
 arcon_error = { version = "0.1", path = "arcon_error" }
 arcon_extra = { version = "0.1", path = "arcon_extra" }
-weld = "0.3.0"
 kompact = "0.8.0"
 futures = "0.1"
 regex = "1"
