+++ execution-plane/arcon/arcon_error/src/lib.rs
@@ -17,17 +17,8 @@ macro_rules! arcon_err_kind {
     })
 }
 
-/// Helper macro for defining Weld errors
-#[macro_export]
-macro_rules! weld_error {
-    ( $($arg:tt)* ) => ({
-        $crate::Error::new_weld(format!($($arg)*))
-    })
-}
-
 #[derive(Debug)]
 pub enum ErrorKind {
-    WeldError(String),
     ArconError(String),
 }
 
@@ -41,7 +32,6 @@ impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let msg = match self.kind {
             ErrorKind::ArconError(ref err) => err,
-            ErrorKind::WeldError(ref err) => err,
         };
         write!(f, "{}", msg)
     }
@@ -59,12 +49,6 @@ impl StdError for Error {
 impl Error {
     pub fn new(kind: ErrorKind) -> Self {
         Self { kind, cause: None }
-    }
-    pub fn new_weld(err_msg: String) -> Self {
-        Self {
-            kind: ErrorKind::WeldError(err_msg),
-            cause: None,
-        }
     }
     pub fn new_arcon(err_msg: String) -> Self {
         Self {
