+++ execution-plane/arcon/src/streaming/mod.rs
@@ -1,5 +1,5 @@
 pub mod channel;
 pub mod sink;
 pub mod source;
-pub mod task;
-pub mod window;
+//pub mod window;
+pub mod node;
