+++ execution-plane/arcon/src/streaming/node/mod.rs
@@ -1,11 +1,46 @@
+pub mod operator;
+
 use crate::prelude::*;
 use std::collections::BTreeMap;
 use std::collections::HashSet; // Blocked-list
 use std::collections::LinkedList; // Message buffer
 use std::mem; // Watermark-list
+use std::cmp::Ordering;
+use operator::Operator;
+
+#[derive(Eq, Hash, Copy, Clone, Debug)]
+pub struct NodeID {
+    pub id: u32,
+}
+
+impl NodeID {
+    pub fn new(new_id: u32) -> NodeID {
+        NodeID { id: new_id }
+    }
+}
+impl From<u32> for NodeID {
+    fn from(id: u32) -> Self {
+        NodeID::new(id)
+    }
+}
+impl Ord for NodeID {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.id.cmp(&other.id)
+    }
+}
+impl PartialOrd for NodeID {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+impl PartialEq for NodeID {
+    fn eq(&self, other: &Self) -> bool {
+        self.id == other.id
+    }
+}
 
 /*
-  Node: contains Task which executes actions
+  Node: contains an Operator which executes actions
 */
 #[derive(ComponentDefinition)]
 pub struct Node<IN, OUT>
@@ -17,7 +52,7 @@ where
     id: NodeID,
     out_channels: Box<ChannelStrategy<OUT>>,
     in_channels: Vec<NodeID>,
-    task: Box<Task<IN, OUT>>,
+    operator: Box<Operator<IN, OUT>>,
     watermarks: BTreeMap<NodeID, Watermark>,
     current_watermark: u64,
     current_epoch: u64,
@@ -34,7 +69,7 @@ where
         id: NodeID,
         in_channels: Vec<NodeID>,
         out_channels: Box<ChannelStrategy<OUT>>,
-        task: Box<Task<IN, OUT>>,
+        operator: Box<Operator<IN, OUT>>,
     ) -> Node<IN, OUT> {
         // Initiate our watermarks
         let mut watermarks = BTreeMap::new();
@@ -47,7 +82,7 @@ where
             id,
             out_channels,
             in_channels,
-            task,
+            operator,
             watermarks,
             current_watermark: 0,
             current_epoch: 0,
@@ -69,7 +104,7 @@ where
 
         match message.event {
             ArconEvent::Element(e) => {
-                let results = self.task.handle_element(e)?;
+                let results = self.operator.handle_element(e)?;
                 for result in results {
                     self.output_event(result)?;
                 }
@@ -100,7 +135,7 @@ where
                     self.current_watermark = new_watermark.timestamp;
 
                     // Handle the watermark
-                    for result in self.task.handle_watermark(new_watermark)? {
+                    for result in self.operator.handle_watermark(new_watermark)? {
                         self.output_event(result)?;
                     }
 
@@ -117,8 +152,8 @@ where
                     // update current epoch
                     self.current_epoch = e.epoch;
 
-                    // call handle_epoch on our task
-                    let _task_state = self.task.handle_epoch(e)?;
+                    // call handle_epoch on our operator
+                    let _operator_state = self.operator.handle_epoch(e)?;
 
                     // store the state
                     self.save_state()?;
@@ -226,8 +261,8 @@ mod tests {
     use super::*;
     use std::sync::Arc;
     use std::{thread, time};
+    use operator::filter::Filter; 
 
-    // Helper functions for cleaner test-cases
     fn node_test_setup() -> (ActorRef<ArconMessage<i32>>, Arc<Component<DebugSink<i32>>>) {
         // Returns a filter Node with input channels: sender1..sender3
         // And a debug sink receiving its results
@@ -237,14 +272,16 @@ mod tests {
         let channel = Channel::Local(sink.actor_ref());
         let channel_strategy: Box<ChannelStrategy<i32>> = Box::new(Forward::new(channel));
 
-        let weld_code = String::from("|x: i32| x >= 0");
-        let module = Arc::new(Module::new(weld_code).unwrap());
+        fn node_fn(x: &i32) -> bool {
+            *x >= 0
+        }
+
         let filter_node = system.create_and_start(move || {
             Node::<i32, i32>::new(
                 0.into(),
                 vec![1.into(), 2.into(), 3.into()],
                 channel_strategy,
-                Box::new(Filter::<i32>::new(module)),
+                Box::new(Filter::new(&node_fn)),
             )
         });
