+++ execution-plane/arcon/src/streaming/node/operator/mod.rs
@@ -1,0 +1,18 @@
+pub mod filter;
+pub mod map;
+
+pub use filter::Filter;
+pub use map::Map;
+
+use crate::data::{ArconElement, ArconEvent, ArconType, Epoch, Watermark};
+use arcon_error::ArconResult;
+
+pub trait Operator<IN, OUT>
+where
+    IN: 'static + ArconType,
+    OUT: 'static + ArconType,
+{
+    fn handle_element(&mut self, element: ArconElement<IN>) -> ArconResult<Vec<ArconEvent<OUT>>>;
+    fn handle_watermark(&mut self, watermark: Watermark) -> ArconResult<Vec<ArconEvent<OUT>>>;
+    fn handle_epoch(&mut self, epoch: Epoch) -> ArconResult<Vec<u8>>;
+}
