--- execution-plane/arcon/src/weld/mod.rs
@@ -1,7 +1,0 @@
-pub mod future;
-pub mod module;
-
-pub use future::*;
-pub use module::{Module, ModuleRun};
-pub use weld_core::data::*;
-pub use weld_core::*;
